requirements.txt
streamlit
langchain
langchain-community
python-dotenv
google-cloud-aiplatform
google-api-python-client
google-auth-oauthlib
google-cloud-texttospeech
google-cloud-speech
spotipy
pygame
googlemaps
pypdf
numpy
google-generativeai
opencv-python
image
tweepy
python-dotenv
# Pillow==9.4.0
Pillow
google-cloud-documentai
google-cloud-discoveryengine
matplotlib
seaborn
beautifulsoup4
streamlit-carousel
flask
langchainhub
vertexai
langchain-google-vertexai
# transformers
# torch 
# torchvision 
# accelerate